statistics.py:  - New file  - LibraryStats and SummaryStats have been removed from FeatureSelector and placed here
FeatureSelector.py:
 - Imports updated to reflect new location for LibraryStats and SummaryStats
 - We are no longer tracking filtered/unknown/uncounted features, which cleaned up the code a bit. FeatureSelector.choose() now returns only a set of selected features rather than a set of hit tuples
 - Readability changes

counter2_0.py:
 - Minor readability changes and imports changes to reflect the new package structure for stats collection classes
